kill sync15-traits
