+++ Cargo.toml
@@ -23,7 +23,6 @@ members = [
     "components/support/rc_crypto/nss/nss_sys",
     "components/support/rc_crypto/nss/systest",
     "components/support/sql",
-    "components/support/sync15-traits",
     "components/support/types",
     "components/support/viaduct-reqwest",
     "components/sync_manager",
@@ -89,7 +88,6 @@ default-members = [
     "components/support/rc_crypto/nss/nss_build_common",
     "components/support/rc_crypto/nss/nss_sys",
     "components/support/sql",
-    "components/support/sync15-traits",
     "components/support/types",
     "components/support/viaduct-reqwest",
     "components/sync_manager",
