--- components/support/sync15-traits/README.md
@@ -1,10 +1,0 @@
-# sync15-traits
-
-The sync15-traits support component is a home for the core types and traits
-used by Sync. It's extracted into its own crate because these types are used
-in so many different contexts, from the [Sync Manager](../sync_manager/index.html)
-to desktop Firefox.
-
-Because this is a "support" component, there aren't any instructions for how
-to use this component, but you will find documentation for most of the types
-and traits implemented here.
