+++ components/sync15/Cargo.toml
@@ -13,7 +13,7 @@ default = ["sync-engine"]
 random-guid = ["sync-guid/random"]
 
 # Some consumers of this just need our encrypted payloads and no other sync functionality.
-crypto = ["rc_crypto", "base16", "base64", "sync15-traits/crypto"]
+crypto = ["rc_crypto", "base16", "base64"]
 
 # Some crates need to implement a "sync engine", but aren't a "sync client" (ie, their
 # engine is used by a "sync client".) Engines don't interact directly with the storage servers,
@@ -41,11 +41,10 @@ interrupt-support = { path = "../support/interrupt" }
 lazy_static = "1.4"
 log = "0.4"
 rc_crypto = { path = "../support/rc_crypto", features = ["hawk"], optional = true }
-serde = "1"
+serde = { version = "1", features = ["derive"] }
 serde_derive = "1"
 serde_json = "1"
 sync-guid = { path = "../support/guid", features = ["random"] }
-sync15-traits = {path = "../support/sync15-traits" }
 thiserror = "1.0"
 url = { version = "2.1", optional = true } # mozilla-central can't yet take 2.2 (see bug 1734538)
 viaduct = { path = "../viaduct", optional = true }
