+++ components/sync15/src/clients/record.rs
@@ -134,7 +134,7 @@ impl From<Command> for CommandRecord {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use sync15_traits::Payload;
+    use crate::Payload;
 
     #[test]
     fn test_ttl() {
