+++ components/sync15/src/key_bundle.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use crate::error::{Result, SyncTraitsError as Error};
+use crate::error::{Error, Result};
 use rc_crypto::{
     aead::{self, OpeningKey, SealingKey},
     rand,
