+++ components/sync15/src/lib.rs
@@ -5,6 +5,8 @@
 #![allow(unknown_lints, clippy::implicit_hasher)]
 #![warn(rust_2018_idioms)]
 
+#[cfg(feature = "crypto")]
+mod bso_record;
 #[cfg(feature = "sync-client")]
 pub mod client;
 // Types to describe client records
@@ -16,7 +18,12 @@ pub mod clients;
 #[cfg(feature = "sync-client")]
 mod collection_keys;
 mod error;
+#[cfg(feature = "crypto")]
+mod key_bundle;
+
+mod payload;
 pub(crate) mod record_types;
+mod server_timestamp;
 
 #[cfg(feature = "sync-engine")]
 pub mod engine;
@@ -27,11 +34,18 @@ pub use crate::client_types::{ClientData, DeviceType, RemoteClient};
 #[cfg(feature = "sync-client")]
 pub use crate::collection_keys::CollectionKeys;
 pub use crate::error::{Error, Result};
-pub use sync15_traits::SyncTraitsError;
 
 pub use sync_guid::Guid;
 
 #[cfg(feature = "crypto")]
-pub use sync15_traits::{BsoRecord, CleartextBso, EncryptedBso, EncryptedPayload, KeyBundle};
+pub use bso_record::{BsoRecord, CleartextBso, EncryptedBso, EncryptedPayload};
+#[cfg(feature = "crypto")]
+pub use key_bundle::KeyBundle;
 
-pub use sync15_traits::{Payload, ServerTimestamp};
+pub use payload::Payload;
+pub use server_timestamp::ServerTimestamp;
+
+// For skip_serializing_if
+pub(crate) fn skip_if_default<T: PartialEq + Default>(v: &T) -> bool {
+    *v == T::default()
+}
